requirements.txt
python-telegram-bot>=21.0,<22
qrcode==7.4.2
Pillow>=10.2,<12
